+++ core-design/src/main/java/com/ddd/ansayo/core_design/util/ConstantValues.kt
@@ -1,0 +1,3 @@
+package com.ddd.ansayo.core_design.util
+
+const val MAX_LENGTH_COURSE_SEARCH= 40
