+++ core-design/src/main/java/com/ddd/ansayo/core_design/util/button/LabelLineButton.kt
@@ -1,4 +1,4 @@
-package com.ddd.ansayo.core_design.util
+package com.ddd.ansayo.core_design.util.button
 
 import android.content.Context
 import android.util.AttributeSet
