+++ core-design/src/main/java/com/ddd/ansayo/core_design/util/toast/ToastLineMax.kt
@@ -1,4 +1,4 @@
-package com.ddd.ansayo.core_design.util
+package com.ddd.ansayo.core_design.util.toast
 
 import android.app.Activity
 import android.content.Context
