+++ core-design/src/main/res/drawable/bg_text_filed_search.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+
+    <solid android:color="@color/N100"/>
+    <corners android:radius="36dp" />
+
+</shape>
