+++ core-design/src/main/res/drawable/bg_text_filed_short.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:drawable="@drawable/text_filed_short_focused" android:state_focused="true" />
+
+    <item android:drawable="@drawable/button_label_secondary_enable" android:state_focused="false" />
+
+</selector>
