+++ core-design/src/main/res/drawable/text_filed_short_error.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+
+    <solid android:color="@color/N90"/>
+    <stroke android:width="1dp" android:color="@color/red_caution" />
+    <corners android:radius="12dp" />
+
+</shape>
