+++ core-design/src/main/res/layout/ds_button_line_default.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="view"
-            type="com.ddd.ansayo.core_design.util.LabelLineButton" />
+            type="com.ddd.ansayo.core_design.util.button.LabelLineButton" />
 
     </data>
