+++ core-design/src/main/res/layout/ds_button_location_add.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="view"
-            type="com.ddd.ansayo.core_design.util.LabelLocationButton" />
+            type="com.ddd.ansayo.core_design.util.button.LabelLocationButton" />
 
     </data>
