+++ core-design/src/main/res/layout/ds_button_primary_bottom_max.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="view"
-            type="com.ddd.ansayo.core_design.util.LabelPrimaryBottomMaxButton" />
+            type="com.ddd.ansayo.core_design.util.button.LabelPrimaryBottomMaxButton" />
 
     </data>
