+++ core-design/src/main/res/layout/ds_button_primary_default.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="view"
-            type="com.ddd.ansayo.core_design.util.LabelPrimaryButton" />
+            type="com.ddd.ansayo.core_design.util.button.LabelPrimaryButton" />
 
     </data>
