+++ core-design/src/main/res/layout/ds_button_secondary_default.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="view"
-            type="com.ddd.ansayo.core_design.util.LabelSecondaryButton" />
+            type="com.ddd.ansayo.core_design.util.button.LabelSecondaryButton" />
 
     </data>
