+++ core-design/src/main/res/layout/ds_button_toggle_default.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="view"
-            type="com.ddd.ansayo.core_design.util.LabelToggleButton" />
+            type="com.ddd.ansayo.core_design.util.button.LabelToggleButton" />
 
     </data>
