+++ core-design/src/main/res/layout/ds_text_filed_search.xml
@@ -1,0 +1,65 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools">
+
+    <data>
+
+        <variable
+            name="view"
+            type="com.ddd.ansayo.core_design.util.textfiled.TextFiledShort" />
+
+        <variable
+            name="label"
+            type="String" />
+
+        <variable
+            name="enabled"
+            type="Boolean" />
+
+        <variable
+            name="hint"
+            type="String" />
+
+        <variable
+            name="text"
+            type="String" />
+
+    </data>
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content">
+
+        <EditText
+            android:id="@+id/et_content"
+            style="@style/Text_B1"
+            android:layout_width="0dp"
+            android:layout_height="48dp"
+            android:enabled="@{enabled}"
+            android:hint="@{hint}"
+            android:textColorHint="@color/N70"
+            android:imeOptions="actionDone"
+            android:inputType="text"
+            android:text="@={text}"
+            android:background="@drawable/bg_text_filed_search"
+            android:textCursorDrawable="@drawable/ds_ggeco_cursor_default"
+            android:paddingStart="16dp"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            tools:hint="장소명이나 주소를 입력해보세요."/>
+
+        <TextView
+            android:id="@+id/tv_length"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:textSize="11dp"
+            android:textColor="@color/N60"
+            android:layout_marginEnd="10dp"
+            app:layout_constraintTop_toTopOf="@id/et_content"
+            app:layout_constraintBottom_toBottomOf="@id/et_content"
+            app:layout_constraintEnd_toEndOf="parent"
+            tools:text="0/0"/>
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
+</layout>
