+++ core-design/src/main/res/values/strings.xml
@@ -1,0 +1,6 @@
+<resources>
+
+    <!--Design System-->
+    <string name="text_filed_length">%1$d / %2$d</string>
+
+</resources>
